backend/requirements.txt
Django>=5.1
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
httpx>=0.27.0
Pillow>=11.0.0
python-dotenv>=1.0.1
